+++ src/common/libs.ts
@@ -6,18 +6,7 @@ import * as chokidar from 'chokidar';
 import * as moment from 'moment';
 import * as minimist from 'minimist';
 import * as inquirer from 'inquirer';
-import { debounce } from 'lodash';
 
-export {
-  R,
-  fs,
-  fsPath,
-  chokidar,
-  jsYaml,
-  moment,
-  debounce,
-  minimist,
-  inquirer,
-};
+export { R, fs, fsPath, chokidar, jsYaml, moment, minimist, inquirer };
 export { Subject, Observable } from 'rxjs';
 export { log } from 'log.server';