+++ src/index.ts
@@ -15,7 +15,6 @@ export {
   Subject,
   Observable,
   moment,
-  debounce,
   listr,
   IListr,
   IListrOptions,